requirements.txt
Flask
pymongo
python-dotenv
werkzeug
yfinance
pandas
requests
scikit-learn
numpy
flask-bcrypt
